requirements.txt
Django
python-decouple
netifaces
psycopg2
whitenoise
django-extensions
wagtail
django-taggit
django-modelcluster
django-axes[ipware]
pre-commit
black
isort
flake8
